+++ DarkCodex/changelog.md
@@ -16,9 +16,7 @@
 - extra rogue talent
 - buffed angel's light
 - cackle/shant passive activatable
-
-### Changed
-- testing out new parameters on mobile gathering
+- ability focus
 
 ## [1.0.1]